Add files via upload
"I used Python and Matplotlib to create this beautiful visualization of the Indian flag. The flag is made up of three vertical stripes of saffron, white, and green, with a Ashoka Chakra in the center. The saffron stripe represents courage, the white stripe represents peace, and the green stripe represents prosperity. The Ashoka Chakra is a Buddhist symbol of good governance and righteousness. I hope you enjoy this visualization of our national flag!"